pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caduzera"
version = "0.1.0"
description = "Libs reutilizáveis do Cadu"
requires-python = ">=3.10"
dependencies = [
  "pydantic-settings>=2.4",         # base (leve)
  "python-dotenv>=1.0",             # carregar .env facilmente
]

[project.optional-dependencies]
telegram = [
  "telethon>=1.36.0",
]

[tool.setuptools.packages.find]
where = ["src"]
